chore: suppress eslint warnings and include editor config for formatting
