+++ .editorconfig
@@ -1,0 +1,23 @@
+# Editor configuration, see https://editorconfig.org
+root = true
+
+[*]
+charset = utf-8
+end_of_line = lf
+indent_size = 2
+indent_style = space
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.md]
+trim_trailing_whitespace = false
+
+[*.js]
+# Change these settings to your own preference
+indent_size = 2
+indent_style = space
+
+[*.ts]
+# Change these settings to your own preference
+indent_size = 2
+indent_style = space
