+++ .eslintrc
@@ -1,15 +1,48 @@
 {
   "root": true,
+  "env": {
+    "es2021": true,
+    "node": true
+  },
   "parser": "@typescript-eslint/parser",
-  "plugins": ["@typescript-eslint"],
+  "parserOptions": {
+    "project": "./tsconfig.json"
+  },
+  "ignorePatterns": ["src/__tests__/*"],
+  "plugins": ["import", "@typescript-eslint", "simple-import-sort"],
   "extends": [
     "eslint:recommended",
+    "plugin:eslint-comments/recommended",
     "plugin:@typescript-eslint/eslint-recommended",
-    "plugin:@typescript-eslint/recommended"
+    "plugin:@typescript-eslint/recommended",
+    "plugin:@typescript-eslint/recommended-requiring-type-checking",
+    "plugin:@typescript-eslint/strict",
+    "plugin:import/recommended",
+    "plugin:import/typescript",
+    "prettier"
   ],
+  "settings": {
+    "import/resolver": {
+      "typescript": {
+        "project": "tsconfig.json"
+      }
+    }
+  },
   "rules": {
-    "no-console": 1,
-    "semi-style": ["error", "last"],
-    "semi": [2, "always"]
+    "prefer-const": "warn",
+    "import/no-mutable-exports": "error",
+    "import/no-unused-modules": "error",
+    "simple-import-sort/imports": "error",
+    "simple-import-sort/exports": "error",
+    "import/no-extraneous-dependencies": "error",
+    "@typescript-eslint/no-unused-vars": "warn",
+    "@typescript-eslint/no-explicit-any": "warn",
+    "@typescript-eslint/no-unsafe-call": "warn",
+    "@typescript-eslint/no-unsafe-return": "warn",
+    "@typescript-eslint/no-unsafe-argument": "off",
+    "@typescript-eslint/no-unsafe-assignment": "off",
+    "@typescript-eslint/no-non-null-assertion": "off",
+    "@typescript-eslint/no-unsafe-member-access": "off",
+    "import/consistent-type-specifier-style": ["error", "prefer-top-level"]
   }
 }
